requirements.txt
requests==2.32.5
python-dotenv==1.2.1
openai==2.20.0
openai-agents[litellm]==0.9.0
beautifulsoup4==4.14.2
mcp[cli]==1.26.0
jupyter-black==0.4.0
httpx[socks]==0.28.1
torch==2.10.0
transformers==5.1.0
opik==1.10.12
numpy==2.4.1
pandas==2.3.3
seaborn==0.13.2
accelerate==1.12.0
nnsight==0.5.15
gdown==5.2.1
typer==0.24.1
pyright==1.1.400
scikit-learn==1.8.0
lightning==2.6.1
datasets==4.5.0
ipywidgets==8.1.8
peft==0.18.1
trl==0.27.2
pydantic==2.12.5
sqlalchemy==2.0.41
pyyaml==6.0.2
psycopg2==2.9.11
asyncpg==0.30.0
greenlet==3.3.2
python-telegram-bot==22.6
